+++ bin/trim.js
@@ -1,7 +1,17 @@
 #!/usr/bin/env node
 
-const commands = require('../src/commands');
+const contentful = require('../src/contentful');
+const orphanedAssetTrimmer = require('../src/orphaned-asset-trimmer');
+const outdatedEntryTrimmer = require('../src/outdated-entry-trimmer');
 const yargs = require('yargs');
+
+const reportError = (error) => {
+    try {
+        error = JSON.parse(error.message);
+    } catch (ignore) {}
+
+    console.error('Error: ' + error.message);
+};
 
 yargs
     .version()
@@ -18,10 +28,30 @@ yargs
         default: 5
     })
     .command('orphaned-assets <space> <token>', 'delete unused assets', {}, async function (argv) {
-        await commands.trimOrphanedAssets(argv.space, argv.token, argv.gracePeriod, argv.dryRun);
+        contentful.config.gracePeriod = argv.gracePeriod;
+        contentful.config.isDryRun = argv.dryRun;
+
+        try {
+            const space = await contentful.getSpace(argv.space, argv.token);
+            const stats = await orphanedAssetTrimmer.trim(space);
+
+            console.log(`Deleted ${stats.deletedCount} orphaned assets.`);
+        } catch (e) {
+            reportError(e);
+        }
     })
     .command('outdated-entries <space> <token> <field>', 'delete entries where <field> is in the past', {}, async function (argv) {
-        await commands.trimOutdatedEntries(argv.space, argv.token, argv.field, argv.gracePeriod, argv.dryRun);
+        contentful.config.gracePeriod = argv.gracePeriod;
+        contentful.config.isDryRun = argv.dryRun;
+
+        try {
+            const space = await contentful.getSpace(argv.space, argv.token);
+            const stats = await outdatedEntryTrimmer.trim(space, argv.field);
+
+            console.log(`Deleted ${stats.deletedCount} outdated entries.`);
+        } catch (e) {
+            reportError(e);
+        }
     })
     .demandCommand(1, 'Please provide a valid command.')
     .help()
